+++ lib/simple_dart_table.dart
@@ -1,11 +1,13 @@
 library simple_dart_table;
 
+export 'src/abstract_table_cell.dart';
+export 'src/abstract_table_row.dart';
+export 'src/cell_renderer_default.dart';
 export 'src/table.dart';
 export 'src/table_cells.dart';
 export 'src/table_column_descr.dart';
-export 'src/table_column_to_header_cell_default.dart';
+export 'src/table_copy.dart';
 export 'src/table_image.dart';
 export 'src/table_link.dart';
 export 'src/table_row.dart';
-export 'src/table_value_to_cell_default.dart';
 export 'src/table_value_to_string_default.dart';