+++ lib/src/table_column_to_header_cell_default.dart
@@ -1,4 +1,0 @@
-import 'table_cells.dart';
-import 'table_column_descr.dart';
-
-AbstractTableCell tableColumnToHeaderCellDefault(TableColumnDescr columnDescr) => ColumnHeaderCell(columnDescr);