+++ lib/src/table_row.dart
@@ -1,19 +1,54 @@
-import 'package:simple_dart_ui_core/simple_dart_ui_core.dart';
+import '../simple_dart_table.dart';
 
-import 'table_cells.dart';
+abstract class CellRenderer {
+  bool checkCellByType(AbstractTableCell cell, dynamic value);
 
-class TableRow extends PanelComponent {
-  List<AbstractTableCell> cells = <AbstractTableCell>[];
+  AbstractTableCell createCellByType(TableColumnDescr columnDescr, dynamic value);
+}
 
-  List<dynamic> get data => cells.map((e) => e.value).toList();
+class TableRow extends AbstractTableRow {
+  late CellRenderer cellFactory;
 
-  set data(List<dynamic> value) {
-    for (var i = 0; i < value.length; i++) {
-      cells[i].value = value[i];
+  final List<AbstractTableCell> _cells = <AbstractTableCell>[];
+
+  @override
+  List<AbstractTableCell> get cells => _cells;
+
+  @override
+  List<dynamic> get rowData => cells.map((e) => e.value).toList();
+
+  @override
+  set rowData(List<dynamic> value) {
+    for (var colNum = 0; colNum < value.length; colNum++) {
+      createOrUpdateCell(colNum, value[colNum]);
     }
   }
 
-  TableRow() : super('TableRow') {
-    fullWidth();
+  TableRow(List<TableColumnDescr> newCols) : super(newCols) {
+    cellFactory = CellRendererDefault();
+  }
+
+  void createOrUpdateCell(int colNum, dynamic value) {
+    final existCell = cells.length > colNum ? cells[colNum] : null;
+    final columnDescr = columns.length > colNum ? columns[colNum] : TableColumnDescr();
+    if (existCell == null) {
+      final cell = cellFactory.createCellByType(columnDescr, value)
+        ..value = value
+        ..width = '${columnDescr.width}px';
+      cells.add(cell);
+      add(cell);
+    } else {
+      final isCompatibleCell = cellFactory.checkCellByType(existCell, value);
+      if (isCompatibleCell) {
+        existCell.value = value;
+      } else {
+        final cell = cellFactory.createCellByType(columnDescr, value)
+          ..value = value
+          ..width = '${columnDescr.width}px';
+        cells[colNum].remove();
+        insert(colNum, cell);
+        cells[colNum] = cell;
+      }
+    }
   }
 }